agents/requirements.txt
langchain==0.1.9
langchain-community==0.0.38
langchain-groq==0.1.3
langgraph==0.0.26
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pgvector==0.2.5
redis==5.0.1
python-dotenv==1.0.1
httpx==0.26.0
sentence-transformers==2.7.0
